Fix: RoomDetail 구조 수정 및 student 배열 추가하는 중
